Fix and refactor consumed samples save/restore for Megatron models. (#5077)
* Fixes and refactor


* Fix


* Remove unused imports


* Empty


* Fix
